adding features that remove all regions

+++ backend/src/database/database.py
@@ -98,8 +98,14 @@ class Database(metaclass=SingletonMeta):
         select count(*) from chunks;"""
         result = self.fetchone(qry)
         count_regions = result[0]
-        print(f"""Database Report:\nThere is {count_tifs} tifs, {count_meshes} Meshes, {count_regions} User Defined Regions\n""")
-    
+        print(f"""Database Report:\nThere is {count_tifs} tifs, {count_meshes} Meshes, {count_regions} User Defined Regions""")
+        qry = """
+        select lat_begin, lat_end, lon_begin, lon_end, filename, origin_lat, origin_lon from tifs order by lat_begin, lon_begin;"""
+        result = self.fetchall(qry)
+        for each in result:
+            print(f"- {each[4]}[{each[5]},{each[6]}]: latitude from {int(each[0])} to {int(each[1])}, longitude from {int(each[2])} to {int(each[3])}.")
+        print("")
+
     def put_cache(self, cache_key, cache_value):
         self.cache.update({cache_key: cache_value})
 
@@ -109,25 +115,71 @@ class Database(metaclass=SingletonMeta):
             return cache_value
         return value
     
+    def remove_cache(self, cache_key):
+        if cache_key is not None:
+            value = self.cache.get(cache_key)
+            if value is not None:
+                qry = """
+                delete from chunks where id = ?;
+                """
+                self.execute_in_worker(qry, [value['id']])
+                mesh_id = value['mesh_id']
+                if mesh_id is not None:
+                    pth = self.fetchone("select pth from meshes where id = ?;", [mesh_id])[0]
+                    self.execute_in_worker("delete from meshes where id = ?;", [mesh_id])
+                    if os.path.exists(pth):
+                        os.remove(pth)
+
+    def delete_resource(self, id):
+        qry = """
+        select id, mesh from chunks where id = ?;
+        """
+        result = self.fetchone(qry, [id])
+        if result is None:
+            return
+        qry = """
+        delete from chunks where id = ?;
+        """
+        self.execute_in_worker(qry, [result[0]])
+        mesh_id = result[1]
+        if mesh_id is not None:
+            pth = self.fetchone("select pth from meshes where id = ?;", [mesh_id])[0]
+            self.execute_in_worker("delete from meshes where id = ?;", [mesh_id])
+            if os.path.exists(pth):
+                os.remove(pth)
+
+        pass     
+    
     def in_cache(self, cache_key):
         return cache_key in self.cache.keys()
     
     def clear_cache(self):
         self.cache = {}
         temp = self.cache
-        for value in temp.values():
-            chunk_id = value['id']
-            qry = """
-            delete from chunks where id = ?;
-            """
-            self.execute_in_worker(qry, [chunk_id])
-            mesh_id = value['mesh_id']
-            pth = self.fetchone("select pth from meshes where id = ?;", [mesh_id])[0]
-            self.execute_in_worker("delete from meshes where id = ?;", [mesh_id])
-            if os.path.exists(pth):
-                os.remove(pth)
+        for key in temp.keys():
+            self.remove_cache(key)
+
+    def clear_regions(self):
+
+        # for value in temp.values():
+        #     chunk_id = value['id']
+        qry = """
+        select id from chunks;
+        """
+        result = self.fetchall(qry)
+        for each in result:
+            self.delete_resource(each[0])
+
+        #     self.execute_in_worker(qry, [chunk_id])
+        #     mesh_id = value['mesh_id']
+        #     pth = self.fetchone("select pth from meshes where id = ?;", [mesh_id])[0]
+        #     self.execute_in_worker("delete from meshes where id = ?;", [mesh_id])
+        #     if os.path.exists(pth):
+        #         os.remove(pth)
                 
         pass
+
+    # def 
 
 tables = """
 create table if not exists meshes (
